Waiver saving/loading on manage waivers page, TODO: front end work, volunteer perspective
