--- src/pages/manage-waivers/edit.jsx
@@ -1,3 +1,0 @@
-import WaiverEdit from "../../screens/Waivers/WaiverEdit";
-
-export default WaiverEdit;
